+++ app/src/main/java/com/marverenic/music/fragments/AlbumFragment.java
@@ -13,8 +13,8 @@ import com.marverenic.music.JockeyApplication;
 import com.marverenic.music.R;
 import com.marverenic.music.data.store.MusicStore;
 import com.marverenic.music.instances.Album;
-import com.marverenic.music.instances.section.AlbumSection;
-import com.marverenic.music.instances.section.LibraryEmptyState;
+import com.marverenic.music.section.AlbumSection;
+import com.marverenic.music.section.LibraryEmptyState;
 import com.marverenic.music.view.BackgroundDecoration;
 import com.marverenic.heterogeneousadapter.HeterogeneousAdapter;
 import com.marverenic.music.view.GridSpacingDecoration;
