+++ app/src/main/java/com/marverenic/music/fragments/QueueFragment.java
@@ -17,9 +17,9 @@ import android.view.ViewGroup;
 import com.marverenic.heterogeneousadapter.DragDropAdapter;
 import com.marverenic.heterogeneousadapter.DragDropDecoration;
 import com.marverenic.music.R;
-import com.marverenic.music.instances.section.LibraryEmptyState;
-import com.marverenic.music.instances.section.QueueSection;
-import com.marverenic.music.instances.section.SpacerSingleton;
+import com.marverenic.music.section.LibraryEmptyState;
+import com.marverenic.music.section.QueueSection;
+import com.marverenic.music.section.SpacerSingleton;
 import com.marverenic.music.player.PlayerController;
 import com.marverenic.music.view.DragBackgroundDecoration;
 import com.marverenic.music.view.DragDividerDecoration;
