+++ app/src/main/java/com/marverenic/music/section/ArtistBioSingleton.java
@@ -1,4 +1,4 @@
-package com.marverenic.music.instances.section;
+package com.marverenic.music.section;
 
 import android.annotation.SuppressLint;
 import android.content.Intent;
