+++ app/src/main/java/com/marverenic/music/section/RelatedArtistSection.java
@@ -1,4 +1,4 @@
-package com.marverenic.music.instances.section;
+package com.marverenic.music.section;
 
 import android.content.Context;
 import android.graphics.Bitmap;
