+++ app/src/main/java/com/marverenic/music/view/QueueAnimator.java
@@ -3,8 +3,8 @@ package com.marverenic.music.view;
 import android.support.v7.widget.DefaultItemAnimator;
 import android.support.v7.widget.RecyclerView.ViewHolder;
 
-import com.marverenic.music.instances.section.QueueSection;
-import com.marverenic.music.instances.section.SpacerSingleton;
+import com.marverenic.music.section.QueueSection;
+import com.marverenic.music.section.SpacerSingleton;
 
 public class QueueAnimator extends DefaultItemAnimator {
